+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/rendering/VisualGridEyes.java
@@ -651,7 +651,7 @@ public class VisualGridEyes implements ISeleniumEyes, IRenderingEyes {
         Map<String, Object> arguments = new HashMap<String, Object>() {{
             put("serializeResources", true);
             put("dontFetchResources", getConfiguration().isDisableBrowserFetching());
-            put("skipResources", skipList);
+            //put("skipResources", skipList);
         }};
 
         String result = EyesSeleniumUtils.runDomScript(logger, webDriver, userAgent, domScript, arguments, pollingScript);