+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [vNext]
 ### Fixed
 - Added setApiKey and setProxy to BatchClose. [Trello 2189](https://trello.com/c/SlHH9Ssb)
+### Updated
+- Use default viewport height value if we get an exception after `driver.getSystemBars()` call. [Trello 2307](https://trello.com/c/8VCtSmfN)
 
 ## [3.186.0] - 2020-11-23
 ### Updated